completed switch to FullApp. no more TestFullApp
